requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
plyer==2.1.0
discord-webhook==1.3.0 